+++ backend/src/main/java/com/jdsnack/JdSnackApplication.java
@@ -1,0 +1,12 @@
+package com.jdsnack;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JdSnackApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(JdSnackApplication.class, args);
+    }
+}